#333: Address review comment: Move functions to test
